Revert to previous more simple and appropriate modularisation and navigation for this project
