+++ app/src/main/java/com/example/studyapp/navigation/StudyAppNavHost.kt
@@ -1,11 +1,22 @@
 package com.example.studyapp.navigation
 
-/**
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import androidx.navigation.toRoute
+import com.example.studyapp.ui.screens.SubtopicScreen
+import com.example.studyapp.ui.screens.SubtopicsScreen
+import com.example.studyapp.ui.screens.TopicsScreen
+import com.example.studyapp.ui.viewmodels.SubtopicViewModel
+import com.example.studyapp.ui.viewmodels.SubtopicsViewModel
+import com.example.studyapp.ui.viewmodels.TopicsViewModel
+
 @Composable
 fun StudyAppNavHost(modifier: Modifier = Modifier) {
     val navController = rememberNavController()
-// TODO: delete or adapt this to the new layout
-/*
     NavHost(
         navController = navController, startDestination = TopicsRoute, modifier = modifier
     ) {
@@ -35,8 +46,7 @@ fun StudyAppNavHost(modifier: Modifier = Modifier) {
                 },
                 navigateBack = navController::popBackStack
             )
+
         }
-
-
+    }
 }
- */
