+++ app/src/main/java/com/example/studyapp/ui/screens/SubtopicScreen.kt
@@ -30,15 +30,13 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.PreviewLightDark
-import androidx.compose.ui.tooling.preview.PreviewScreenSizes
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil.compose.AsyncImage
 import com.example.studyapp.R
 import com.example.studyapp.data.Subtopic
 import com.example.studyapp.ui.components.SubtopicFullScreenDialog
-import com.example.studyapp.ui.theme.StudyAppTheme
 import com.example.studyapp.ui.viewmodels.SubtopicViewModel
 
 @Composable
@@ -212,23 +210,20 @@ private fun MoreActionsMenu(
     }
 }
 
-@PreviewLightDark
-@PreviewScreenSizes
 @Composable
+@Preview(showSystemUi = true)
 private fun SubtopicScreenPreview() {
-    StudyAppTheme {
-        SubtopicScaffold(
-            subtopic = Subtopic(
-                id = 1,
-                title = "Subtopic Title",
-                description = "Subtopic Description",
-                checked = false,
-                topicId = 1,
-                imageUri = null
-            ),
-            updateSubtopic = {},
-            deleteSubtopic = {},
-            navigateBack = {}
-        )
-    }
+    SubtopicScaffold(
+        subtopic = Subtopic(
+            id = 1,
+            title = "Subtopic Title",
+            description = "Subtopic Description",
+            checked = false,
+            topicId = 1,
+            imageUri = null
+        ),
+        updateSubtopic = {},
+        deleteSubtopic = {},
+        navigateBack = {}
+    )
 }
