--- app/src/main/java/com/example/studyapp/ui/screens/TopicsListDetailScreen.kt
@@ -1,510 +1,0 @@
-package com.example.studyapp.ui.screens
-
-import androidx.activity.compose.BackHandler
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Box
-import androidx.compose.material3.adaptive.ExperimentalMaterial3AdaptiveApi
-import androidx.compose.material3.adaptive.WindowAdaptiveInfo
-import androidx.compose.material3.adaptive.currentWindowAdaptiveInfo
-import androidx.compose.material3.adaptive.layout.AnimatedPane
-import androidx.compose.material3.adaptive.layout.ListDetailPaneScaffoldRole
-import androidx.compose.material3.adaptive.layout.PaneAdaptedValue
-import androidx.compose.material3.adaptive.layout.PaneExpansionAnchor
-import androidx.compose.material3.adaptive.layout.ThreePaneScaffoldDestinationItem
-import androidx.compose.material3.adaptive.layout.calculatePaneScaffoldDirective
-import androidx.compose.material3.adaptive.layout.rememberPaneExpansionState
-import androidx.compose.material3.adaptive.navigation.BackNavigationBehavior
-import androidx.compose.material3.adaptive.navigation.NavigableListDetailPaneScaffold
-import androidx.compose.material3.adaptive.navigation.ThreePaneScaffoldNavigator
-import androidx.compose.material3.adaptive.navigation.ThreePaneScaffoldPredictiveBackHandler
-import androidx.compose.material3.adaptive.navigation.rememberListDetailPaneScaffoldNavigator
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clipToBounds
-import androidx.compose.ui.layout.layout
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import com.example.studyapp.navigation.TopicPlaceholderRoute
-import com.example.studyapp.navigation.TopicRoute
-import com.example.studyapp.navigation.TopicsRoute
-import com.example.studyapp.ui.viewmodels.TopicViewModel
-import com.example.studyapp.ui.viewmodels.Topics2PaneViewModel
-import kotlinx.coroutines.launch
-import kotlin.math.max
-
-
-fun NavGraphBuilder.topicsListDetailScreen() {
-    composable<TopicsRoute> {
-        TopicsListDetailScreen()
-    }
-}
-
-@Composable
-internal fun TopicsListDetailScreen(
-    viewModel: Topics2PaneViewModel = hiltViewModel(),
-    windowAdaptiveInfo: WindowAdaptiveInfo = currentWindowAdaptiveInfo(),
-) {
-    val selectedTopicId by viewModel.selectedTopicId.collectAsStateWithLifecycle()
-    TopicsListDetailScreen(
-        selectedTopicId = selectedTopicId,
-        onTopicClick = viewModel::onTopicClick,
-        windowAdaptiveInfo = windowAdaptiveInfo,
-    )
-}
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-@Composable
-internal fun TopicsListDetailScreen(
-    selectedTopicId: Int?,
-    onTopicClick: (Int) -> Unit,
-    windowAdaptiveInfo: WindowAdaptiveInfo,
-) {
-    val listDetailNavigator = rememberListDetailPaneScaffoldNavigator(
-        scaffoldDirective = calculatePaneScaffoldDirective(windowAdaptiveInfo),
-        initialDestinationHistory = listOfNotNull(
-            ThreePaneScaffoldDestinationItem(ListDetailPaneScaffoldRole.List),
-            ThreePaneScaffoldDestinationItem<Nothing>(ListDetailPaneScaffoldRole.Detail).takeIf {
-                selectedTopicId != null
-            },
-        ),
-    )
-    val coroutineScope = rememberCoroutineScope()
-
-    val paneExpansionState = rememberPaneExpansionState(
-        anchors = listOf(
-            PaneExpansionAnchor.Proportion(0f),
-            PaneExpansionAnchor.Proportion(0.5f),
-            PaneExpansionAnchor.Proportion(1f),
-        ),
-    )
-
-    ThreePaneScaffoldPredictiveBackHandler(
-        listDetailNavigator,
-        BackNavigationBehavior.PopUntilScaffoldValueChange,
-    )
-    BackHandler(
-        paneExpansionState.currentAnchor == PaneExpansionAnchor.Proportion(0f) &&
-                listDetailNavigator.isListPaneVisible() &&
-                listDetailNavigator.isDetailPaneVisible(),
-    ) {
-        coroutineScope.launch {
-            paneExpansionState.animateTo(PaneExpansionAnchor.Proportion(1f))
-        }
-    }
-
-    var topicRoute by remember {
-        val route = selectedTopicId?.let { TopicRoute(id = it) } ?: TopicPlaceholderRoute
-        mutableStateOf(route)
-    }
-
-    fun onTopicClickShowDetailPane(topicId: Int) {
-        onTopicClick(topicId)
-        topicRoute = TopicRoute(id = topicId)
-        coroutineScope.launch {
-            listDetailNavigator.navigateTo(ListDetailPaneScaffoldRole.Detail)
-        }
-        if (paneExpansionState.currentAnchor == PaneExpansionAnchor.Proportion(1f)) {
-            coroutineScope.launch {
-                paneExpansionState.animateTo(PaneExpansionAnchor.Proportion(0f))
-            }
-        }
-    }
-
-    remember { MutableInteractionSource() }
-    val minPaneWidth = 300.dp
-
-    NavigableListDetailPaneScaffold(
-        navigator = listDetailNavigator,
-        listPane = {
-            AnimatedPane {
-                Box(
-                    modifier = Modifier
-                        .clipToBounds()
-                        .layout { measurable, constraints ->
-                            val width = max(minPaneWidth.roundToPx(), constraints.maxWidth)
-                            val placeable = measurable.measure(
-                                constraints.copy(
-                                    minWidth = minPaneWidth.roundToPx(),
-                                    maxWidth = width,
-                                ),
-                            )
-                            layout(constraints.maxWidth, placeable.height) {
-                                placeable.placeRelative(
-                                    x = 0,
-                                    y = 0,
-                                )
-                            }
-                        },
-                ) {
-                    TopicsRoute(
-                        onTopicClick = ::onTopicClickShowDetailPane,
-                        shouldHighlightSelectedTopic = listDetailNavigator.isDetailPaneVisible(),
-                    )
-                }
-            }
-        },
-        detailPane = {
-            AnimatedPane {
-                Box(
-                    modifier = Modifier
-                        .clipToBounds()
-                        .layout { measurable, constraints ->
-                            val width = max(minPaneWidth.roundToPx(), constraints.maxWidth)
-                            val placeable = measurable.measure(
-                                constraints.copy(
-                                    minWidth = minPaneWidth.roundToPx(),
-                                    maxWidth = width,
-                                ),
-                            )
-                            layout(constraints.maxWidth, placeable.height) {
-                                placeable.placeRelative(
-                                    x = constraints.maxWidth -
-                                            max(constraints.maxWidth, placeable.width),
-                                    y = 0,
-                                )
-                            }
-                        },
-                ) {
-                    AnimatedContent(topicRoute) { route ->
-                        when (route) {
-                            is TopicRoute -> {
-                                TopicScreen(
-                                    showBackButton = !listDetailNavigator.isListPaneVisible(),
-                                    onBackClick = {
-                                        coroutineScope.launch {
-                                            listDetailNavigator.navigateBack()
-                                        }
-                                    },
-                                    onTopicClick = ::onTopicClickShowDetailPane,
-                                    viewModel = hiltViewModel<TopicViewModel, TopicViewModel.Factory>(
-                                        key = route.id.toString(),
-                                    ) { factory ->
-                                        factory.create(route.id)
-                                    },
-                                )
-                            }
-
-                            is TopicPlaceholderRoute -> {
-                                TopicDetailPlaceholder()
-                            }
-                        }
-                    }
-                }
-            }
-        },
-        paneExpansionState = paneExpansionState,
-
-        )
-}
-
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-private fun <T> ThreePaneScaffoldNavigator<T>.isListPaneVisible(): Boolean =
-    scaffoldValue[ListDetailPaneScaffoldRole.List] == PaneAdaptedValue.Expanded
-
-@OptIn(ExperimentalMaterial3AdaptiveApi::class)
-private fun <T> ThreePaneScaffoldNavigator<T>.isDetailPaneVisible(): Boolean =
-    scaffoldValue[ListDetailPaneScaffoldRole.Detail] == PaneAdaptedValue.Expanded
-
-
-/*
-@OptIn(ExperimentalMaterial3Api::class, ExperimentalMaterial3AdaptiveApi::class)
-@Composable
-internal fun TopicsScreen(
-    viewModel: TopicsViewModel = hiltViewModel(),
-) {
-    val selectedTopicId by viewModel.selectedSubtopicId.collectAsStateWithLifecycle()
-    val navigationType: StudyAppContentType
-    val contentType: StudyAppContentType
-    var showSearchBar by rememberSaveable { mutableStateOf(false) }
-    val topics by viewModel.topics.collectAsStateWithLifecycle()
-
-    val windowAdaptiveInfo = currentWindowAdaptiveInfo()
-    val listDetailNavigator = rememberListDetailPaneScaffoldNavigator(
-        scaffoldDirective = calculatePaneScaffoldDirective(windowAdaptiveInfo),
-        initialDestinationHistory = listOfNotNull(
-            ThreePaneScaffoldDestinationItem(ListDetailPaneScaffoldRole.List),
-            ThreePaneScaffoldDestinationItem<Nothing>(ListDetailPaneScaffoldRole.Detail).takeIf {
-                selectedTopicId != null
-            },
-        ),
-    )
-
-    if (showSearchBar) {
-        TopicsSearchBar(
-            modifier = modifier,
-            navigateToTopic = navigateToTopic,
-            closeSearchBar = { showSearchBar = false },
-            topics = topics
-        )
-    }
-    TopicsScaffold(
-        topics = topics,
-        createTopic = topicsViewModel::saveTopic,
-        navigateToTopic = navigateToTopic,
-        openSearchBar = { showSearchBar = true },
-        modifier = modifier
-    )
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun TopicsScaffold(
-    topics: List<Topic>?,
-    createTopic: (Topic) -> Unit,
-    navigateToTopic: (Int) -> Unit,
-    openSearchBar: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Scaffold(
-        modifier = modifier,
-        topBar = {
-            CenterAlignedTopAppBar(
-                navigationIcon = {
-                    Icon(
-                        painter = painterResource(R.drawable.baseline_search_24),
-                        contentDescription = stringResource(R.string.topics_search),
-                        modifier = Modifier.clickable { openSearchBar() }
-                    )
-                },
-                actions = {
-                    Icon(
-                        painter = painterResource(R.drawable.baseline_account_circle_24),
-                        contentDescription = stringResource(R.string.go_to_login),
-                        modifier = Modifier.clickable { openSearchBar() }
-                    )
-                },
-                title = { Text(text = stringResource(R.string.app_name)) },
-                modifier = Modifier.padding(horizontal = 16.dp)
-            )
-        },
-        floatingActionButton = {
-            CreateTopicFAB(saveTopic = createTopic)
-        },
-        floatingActionButtonPosition = FabPosition.Center
-    ) { innerPadding ->
-        if (topics == null) {
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                CircularProgressIndicator()
-            }
-        } else {
-            LazyColumn(
-                modifier = Modifier
-                    .padding(innerPadding)
-                    .padding(horizontal = 8.dp)
-            ) {
-                items(topics.size) { index ->
-                    val topic = topics[index]
-                    TopicListItem(
-                        topic = topic,
-                        modifier = Modifier
-                            .clickable { navigateToTopic(topic.id) }
-                            .padding(horizontal = 8.dp)
-                    )
-                }
-            }
-        }
-    }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun TopicsSearchBar(
-    modifier: Modifier = Modifier,
-    navigateToTopic: (Int) -> Unit,
-    topics: List<Topic>?,
-    closeSearchBar: () -> Unit
-) {
-    var query by rememberSaveable { mutableStateOf("") }
-    var filteredTopics by rememberSaveable { mutableStateOf(topics) }
-    SearchBar(
-        inputField = {
-            SearchBarDefaults.InputField(
-                query = query,
-                onQueryChange = { query = it },
-                onSearch = {
-                    filteredTopics = topics?.filter { topic ->
-                        topic.title.contains(query, ignoreCase = true)
-                    }
-                },
-                expanded = true,
-                onExpandedChange = {},
-                leadingIcon = {
-                    Icon(
-                        painter = painterResource(R.drawable.baseline_arrow_back_24),
-                        contentDescription = stringResource(R.string.close_search),
-                        tint = MaterialTheme.colorScheme.onSurface,
-                        modifier = Modifier
-                            .size(24.dp)
-                            .clickable { closeSearchBar() }
-                    )
-                }
-            )
-        },
-        expanded = true,
-        onExpandedChange = {},
-        modifier = modifier,
-        content = {
-            if (topics == null) {
-                Box(
-                    modifier = modifier.fillMaxSize(),
-                    contentAlignment = Alignment.Center
-                ) {
-                    CircularProgressIndicator()
-                }
-            } else {
-                LazyColumn {
-                    items(topics.size) { index ->
-                        val topic = topics[index]
-                        TopicListItem(
-                            topic = topic,
-                            modifier = Modifier.clickable { navigateToTopic(topic.id) }
-                        )
-                    }
-                }
-            }
-        }
-    )
-}
-
-@Composable
-private fun TopicListItem(
-    topic: Topic,
-    modifier: Modifier
-) {
-    ListItem(
-        headlineContent = { Text(topic.title, overflow = TextOverflow.Ellipsis, maxLines = 1) },
-        modifier = modifier,
-        trailingContent = {
-            Checkbox(
-                checked = topic.checked,
-                enabled = false,
-                onCheckedChange = null,
-                modifier = Modifier
-                    .padding(8.dp)
-                    .size(size = 24.dp)
-            )
-        }
-    )
-}
-
-@Composable
-private fun CreateTopicFAB(saveTopic: (Topic) -> Unit, modifier: Modifier = Modifier) {
-    var showDialog by rememberSaveable { mutableStateOf(false) }
-    ExtendedFloatingActionButton(
-        onClick = { showDialog = true },
-        modifier = modifier,
-        icon =
-            {
-                Icon(
-                    painter = painterResource(R.drawable.baseline_add_24),
-                    tint = MaterialTheme.colorScheme.onPrimaryContainer,
-                    contentDescription = stringResource(R.string.create_topic),
-                    modifier = Modifier.size(24.dp)
-                )
-            },
-        text = {
-            Text(
-                text = stringResource(R.string.create_topic),
-                color = MaterialTheme.colorScheme.onPrimaryContainer
-            )
-        },
-    )
-    if (showDialog) {
-        CreateTopicDialog(
-            onDismiss = { showDialog = false },
-            topic = null,
-            onSave = saveTopic
-        )
-    }
-}
-
-@Composable
-fun CreateTopicDialog(
-    topic: Topic?,
-    modifier: Modifier = Modifier,
-    onDismiss: () -> Unit,
-    onSave: (Topic) -> Unit
-) {
-    var topicTitle by rememberSaveable { mutableStateOf(topic?.title ?: "") }
-
-    AlertDialog(
-        onDismissRequest = { onDismiss() },
-        title = { Text(stringResource(R.string.create_topic)) },
-        text = {
-            OutlinedTextField(
-                value = topicTitle,
-                onValueChange = { topicTitle = it },
-                label = { Text(stringResource(R.string.title)) }
-            )
-        },
-        confirmButton = {
-            TextButton(
-                onClick = {
-                    if (topic != null) {
-                        onSave(topic.copy(title = topicTitle))
-                    } else {
-                        onSave(Topic(title = topicTitle, checked = false))
-                    }
-                    onDismiss()
-                }
-            ) {
-                Text(stringResource(R.string.save))
-            }
-        },
-        dismissButton = {
-            TextButton(
-                onClick = { onDismiss() }
-            ) {
-                Text(stringResource(R.string.cancel))
-            }
-        },
-        modifier = modifier
-    )
-}
-
-
-@PreviewLightDark
-@PreviewScreenSizes
-@Composable
-private fun TopicsScreenPreviewLightDark() {
-    StudyAppTheme {
-        TopicsScaffold(
-            topics = listOf(
-                Topic(1, "Topic 1", false),
-                Topic(2, "Topic 2", true),
-                Topic(3, "Topic 3", false)
-            ),
-            navigateToTopic = {},
-            openSearchBar = {},
-            createTopic = {},
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun LoadingScreenPreview() {
-    StudyAppTheme {
-        TopicsScaffold(
-            topics = null,
-            navigateToTopic = {},
-            openSearchBar = {},
-            createTopic = {},
-        )
-    }
-}
-*/
